+++ packages/voidjs-cli/src/ProdBuilder/ClientJsCompiler.ts
@@ -24,7 +24,7 @@ import TerserJSPlugin from 'terser-webpack-plugin'
 import HtmlWebpackPlugin from 'html-webpack-plugin'
 import MiniCssExtractPlugin from 'mini-css-extract-plugin'
 import CssoWebpackPlugin from 'csso-webpack-plugin'
-import WebpackAssetsMap from '../plugins/webpack-assets-map/src/index'
+import WebpackAssetsMap from '../webpack-plugins/webpack-assets-map/src/index'
 import { extensions, alias, rules, cwd } from '../config'
 import prettier from 'prettier'
