+++ packages/voidjs-cli/src/webpack-plugins/HtmlPlugin.ts
