+++ packages/voidjs-cli/src/webpack-plugins/webpack-assets-map/src/index.ts
